requirements.txt
packaging==24.2
PyQt6==6.8.1
PyQt6-3D==6.8.0
PyQt6-Charts==6.8.0
PyQt6-DataVisualization==6.8.0
PyQt6-NetworkAuth==6.8.0
PyQt6-WebEngine==6.8.0
PyQt6_sip==13.10.0
wheel @ file:///usr/local/Cellar/python%403.13/3.13.2/libexec/wheel-0.45.1-py3-none-any.whl#sha256=b9235939e2096903717cb6bfc132267f8a7e46deb2ec3ef9c5e234ea301795d0


# for macOS
# pip3 freeze > requirements.txt
# pip3 install flask, pandas, azure.storage.blob, dotenv
# python3 app.py
Flask
pandas
azure-storage-blob
python-dotenv
gunicorn
